+++ NEWS.md
@@ -1,3 +1,17 @@
+# manymome.table 0.3.2
+
+- The optional arguments (`...`) of
+  `as_flextable.cond_indirect_effects()`
+  and `as_flextable.indirect_list()`
+  are now passed to `flextable::autofit()`
+  when preparing the final output.
+  (0.3.1)
+
+- Updated to print SE-based confidence
+  intervals and *p*-values when a path
+  has no mediator and effect not
+  standardized. (0.3.2)
+
 # manymome.table 0.3.0
 
 - Minor changes to the `README.md`